log-in-master/requirement.txt
flask==1.1.2
Werkzeug==1.0.1
flask-mysql
uuid==1.30
pyjwt==2.0.1
flask-cors==3.0.10